feat: Implement visual feedback for item pickup
When the player picks up an item (e.g., damage upgrade, armor), the name
of the item will be displayed on the screen using the `write()` function.

Side note: The name of all the varibles related to the `Player` class that
included the word 'power' were changed to 'damage' to better illustrate
their purpose.